Start of shop
You can enter the shop area and press e to open the shop and it'll pause the game. It is already pause menu friendly. Opening the pause menu while in the shop will keep the game paused and will close the shop gui, the player will need to reopen the shop if they want to use it still.